requirements.txt
absl-py==0.9.0
astor==0.8.1
backcall==0.2.0
Box2D==2.3.10
cached-property==1.5.1
certifi==2020.4.5.2
cffi==1.14.0
click==7.1.2
cloudpickle==1.3.0
cycler==0.10.0
Cython==0.29.20
decorator==4.4.2
fasteners==0.15
Flask==1.0.2
future==0.18.2
gast==0.2.2
glfw==1.11.2
google-pasta==0.2.0
grpcio==1.29.0
-e git+git@github.com:openai/gym.git@ff4664bd24fa95bbfa1f7682c0723f88a5c6aa3b#egg=gym
h5py==2.10.0
imageio==2.8.0
importlib-metadata==1.6.1
ipython==7.15.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
joblib==0.15.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Lasagne @ https://github.com/Lasagne/Lasagne/archive/master.zip
Mako==1.1.3
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==2.0.2
mkl-service==2.3.0
monotonic==1.5
mujoco-py==2.0.2.10
numpy==1.16.0
opt-einsum==3.2.1
parso==0.7.0
path==14.0.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
plotly==4.0.0
prompt-toolkit==3.0.5
protobuf==3.12.2
ptyprocess==0.6.0
pycparser==2.20
pygame==1.9.6
pyglet==1.5.0
Pygments==2.6.1
PyOpenGL==3.1.5
pyparsing==2.4.7
PyPrind==2.11.2
python-dateutil==2.8.1
pytz==2020.1
retrying==1.3.3
-e git+git@github.com:alexlioralexli/rllab-finetuning.git@2dae9141d0fdc284d04f18931907131d66b43023#egg=rllab
scipy==1.4.1
six==1.15.0
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.2
termcolor==1.1.0
Theano==1.0.4
tqdm==4.46.1
traitlets==4.3.3
-e git+https://github.com/vitchyr/viskit@a0384d69cfdcde8d08646638ce94a13bf1af44d6#egg=viskit
wcwidth==0.2.4
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
